+++ lib/file.cpp
@@ -241,7 +241,7 @@ void File::write(char *buffer, size_t n)
 	fstream.write(buffer, n);
 }
 
-void File::write(char *buffer, size_t n, size_t pos)
+void File::write(char *buffer, size_t n, off_t pos)
 {
 	fstream.seekp(pos + startofdata);
 	fstream.write(buffer, n);
@@ -252,7 +252,7 @@ void File::write(float *buffer, size_t n)
 	fstream.write(reinterpret_cast<char *> (buffer), n*sizeof(float));
 }
 
-void File::write(float *buffer, size_t n, size_t pos)
+void File::write(float *buffer, size_t n, off_t pos)
 {
 	fstream.seekp(pos + startofdata);
 	fstream.write(reinterpret_cast<char *> (buffer), n*sizeof(float));
@@ -263,7 +263,7 @@ void File::write(double *buffer, size_t n)
 	fstream.write(reinterpret_cast<char *> (buffer), n*sizeof(double));
 }
 
-void File::write(double *buffer, size_t n, size_t pos)
+void File::write(double *buffer, size_t n, off_t pos)
 {
 	fstream.seekp(pos + startofdata);
 	fstream.write(reinterpret_cast<char *> (buffer), n*sizeof(double));
@@ -274,7 +274,7 @@ void File::write(int *buffer, size_t n)
 	fstream.write(reinterpret_cast<char *> (buffer), n*sizeof(int));
 }
 
-void File::write(int *buffer, size_t n, size_t pos)
+void File::write(int *buffer, size_t n, off_t pos)
 {
 	fstream.seekp(pos + startofdata);
 	fstream.write(reinterpret_cast<char *> (buffer), n*sizeof(int));
@@ -285,7 +285,7 @@ void File::write(size_t *buffer, size_t n)
 	fstream.write(reinterpret_cast<char *> (buffer), n*sizeof(size_t));
 }
 
-void File::write(size_t *buffer, size_t n, size_t pos)
+void File::write(size_t *buffer, size_t n, off_t pos)
 {
 	fstream.seekp(pos + startofdata);
 	fstream.write(reinterpret_cast<char *> (buffer), n*sizeof(size_t));
@@ -297,7 +297,7 @@ void File::read(char *buffer, size_t n)
 	fstream.read(buffer, n);
 }
 
-void File::read(char *buffer, size_t n, size_t pos)
+void File::read(char *buffer, size_t n, off_t pos)
 {
 	fstream.seekg(pos + startofdata);
 	fstream.read(buffer, n);
@@ -308,7 +308,7 @@ void File::read(float *buffer, size_t n)
 	fstream.read(reinterpret_cast<char *> (buffer), n*sizeof(float));
 }
 
-void File::read(float *buffer, size_t n, size_t pos)
+void File::read(float *buffer, size_t n, off_t pos)
 {
 	fstream.seekg(pos + startofdata);
 	fstream.read(reinterpret_cast<char *> (buffer), n*sizeof(float));
@@ -319,7 +319,7 @@ void File::read(double *buffer, size_t n)
 	fstream.read(reinterpret_cast<char *> (buffer), n*sizeof(double));
 }
 
-void File::read(double *buffer, size_t n, size_t pos)
+void File::read(double *buffer, size_t n, off_t pos)
 {
 	fstream.seekg(pos + startofdata);
 	fstream.read(reinterpret_cast<char *> (buffer), n*sizeof(double));
@@ -330,7 +330,7 @@ void File::read(int *buffer, size_t n)
 	fstream.read(reinterpret_cast<char *> (buffer), n*sizeof(int));
 }
 
-void File::read(int *buffer, size_t n, size_t pos)
+void File::read(int *buffer, size_t n, off_t pos)
 {
 	fstream.seekg(pos + startofdata);
 	fstream.read(reinterpret_cast<char *> (buffer), n*sizeof(int));
@@ -341,7 +341,7 @@ void File::read(size_t *buffer, size_t n)
 	fstream.read(reinterpret_cast<char *> (buffer), n*sizeof(size_t));
 }
 
-void File::read(size_t *buffer, size_t n, size_t pos)
+void File::read(size_t *buffer, size_t n, off_t pos)
 {
 	fstream.seekg(pos + startofdata);
 	fstream.read(reinterpret_cast<char *> (buffer), n*sizeof(size_t));
